+++ src/UChallenge.MockDB/FeiraLivreAggregates/FeiraLivreRepository.cs
@@ -1,7 +1,7 @@
 ﻿namespace UChallenge.MockDB.FeiraLivreAggregates
 {
     public sealed class FeiraLivreRepository :
-        BaseRepository<Domain.FeiraLivreAggregates.FeiraLivre, FeiraLivre, long>,
+        BaseRepository<Domain.FeiraLivreAggregates.FeiraLivre, FeiraLivre, int>,
         Domain.FeiraLivreAggregates.IFeiraLivreRepository
     {
         public FeiraLivreRepository(MockDbContext context) :
